Rename "Administration" to "Board"
Initially, we referred to it as the admin group; however, this terminology can be confusing for those who are not familiar with the PHP Foundation's background.

Furthermore, when delivering talks or communicating with external parties, describing one's title as "PHP Foundation Administrator" can create a misleading impression, as the term is often used differently in the industry.

Additionally, I have added "Operations Manager" to my title because JetBrains currently sponsors me to work on the foundation almost full-time, and I'm handling any inquiries that come to the foundation.