Remove incorrect note for CVE-2024-8928
This was a mix up in the description that was copied from OSTIF document
which took description from another issue:

A memory-related vulnerability in PHP's filter handling system,
particularly when processing input with convert.quoted-printable-decode
filters, leads to a segmentation fault. This vulnerability is triggered
through specific sequences of input data, causing PHP to crash. When
exploited, it allows an attacker to extract a single byte of data from
the heap or cause a DoS. (CVE-2024-11233)

This is taken from ALAS advisory:
https://alas.aws.amazon.com/AL2023/ALAS-2025-845.html

It clearly shows that it is a mix up with CVE-2024-11233. The
CVE-2024-8928 is a different issue though.